Deleted unnecessary files and notebooks
